[BP][2/n] MB-25920: Implement pause-resume capability in StaleItemDeleter
EphTombstoneStaleItemDeleter task can have very long runtimes at both
on average and at tail end. The issue with that is not necessarily the
duration of the task itself; but it will block the other tasks in the
same class (NonIO) from running.

EphTombstoneStaleItemDeleter task does purge across all vbuckets.
When the task runs for a long time and we decide to pause-resume,
we must ensure that resume happens from the vbucket where we had
paused. This task uses the 'PauseResumeVBVisitor' and
'ProgressTracker' to accomplish the pause-resume.

Also unit test that simulates the pause-resume of the task is added.

Change-Id: I6eda0445931e8cf36c0d4b293a4f8d10e5f698f6
Reviewed-on: http://review.couchbase.org/84330
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>