MB-21647: Fix intermittent failure in DCP producer testing
In some DCP producer tests a sporadic failure was observed due to
race condition in waiting for 'last_sent_seqno' stat. That is the
'last_sent_seqno' stat was updated correctly at stream level, but
DCP test client assumed that the item was read immediately with the
updation of the stat, but unfortunately in some cases it was not so.

This commit makes sure that the DCP client explicitly waits for the
last item to be read by the client.

Change-Id: I41d36cf9c4175ecf17db6ba0bed6a885b212ddfe
Reviewed-on: http://review.couchbase.org/84246
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>