Revert 'Update total items using item count from the underlying KVStore'
This partially reverts commit 26f28e2148bbc5137ca423c73541d4430b39472a.

In March 2016, a change was made to how we track vBucket item count
for Full Eviction, to support ForestDB. Prior to this change, we
looked at the result of mutations (create vs. update) to determine if
we should increment the total document count. Similary for deletes; we
decremented the count.

However ForestDB doesn't support returning such status - it cannot
distinguish between create & update in the callback. To allow total
item count to work in this situation, the logic was changed to read
from the underlying KVStore the absolute item count, and use that
value.

Given we are not using ForestDB anymore, revert this change to
simplify item tracking logic and solve some race conditions on
updating numTotalItems which caused it to go negative temporarily.

Change-Id: I99dc5039f2dc57a6798b6d254913db90e82c1c16
Reviewed-on: http://review.couchbase.org/84369
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>