MB-26256: Fix sporadic failure in 'ReplicateJustBeforeThrottleThreshold'
This test relies on memory usage being near the replication threshold.
We were writing a mutation only once to try and get mem_usage above
the threshold, but sometimes mem usage can go down and writing just
one item will not be enough to push it over the threshold.

In this commit, we fix that by wrting multiple mutations until
mem usage goes above the threshold.

Change-Id: I1c7e9da9a4f42a4bf21f33877f0326ba928f526f
Reviewed-on: http://review.couchbase.org/84375
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>