MB-26255[4]: Add stat for total uncompressed data size
Add a stat for the total uncompressed data size. This would be the total
number of bytes that would be sent to the consumer if the data is sent as
uncompressed. This is different from the total bytes sent stat which could
be a lot less if compression is enabled

Change-Id: I50328ff93a034852c19cb9330ba8cc56622fc394
Reviewed-on: http://review.couchbase.org/84084
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>