MB-26218: Don't generate NMVB response packet from the engine
This duplicates the logic for the content of the NMVB packet
being created. Instead the engine should return the standard
error code, and the core will correctly generate the error
message in the standard location.

The problem this patch fix was that as part of refactoring
out the method to generate the NMVB would format and move
the response into the send buffer, except that due to the
return code from the the underlying engine the core thought
that the engine had modified the send buffer.
This buffer turned out to be empty (since it was
already "moved"), and the core would skip sending the
data.

Change-Id: Ib36d86548f718ff42ded69742436995ab9679758
Reviewed-on: http://review.couchbase.org/84503
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>