MB-26436: Reduce list 'readRange' during purge as we iterate over
During purge, we set the 'readRange' from start to end. As the purger
iterates through the list, we don't shorten the 'readRange' by
incrementing the starting point. Note, 'readRange' is intended to
indicate the list update() to mark the item stale and append the new
item value to the end of the list.

This commit, continuously shortens the 'readRange' as the purge
progresses, thereby reducing the number of stale items created by
updates during the purge.

Also, a unit test is added to verify the same.

Change-Id: I52fe6e25b6fe2ecfdd050fcbae59d3798d908e80
Reviewed-on: http://review.couchbase.org/84553
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>