MB-266567: Don't over-count VBucket item counts
TODO: Check what happens in FE if an item is created and deleted in
the same flusher step - I believe the create will be de-duplicated
with the delete; so only the delete persistence callback will run. In
that case I think the item count may be incorrect - or at the very
least it may require newCacheItem to handle it correctly.

Change-Id: I20b53b2c475c75382b84ecad434cf8eabb891378
Reviewed-on: http://review.couchbase.org/84881
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>