MB-26431 [RocksDB] Fix 'ep_testsuite' tests
The 'RocksDBKVStore::getSeqnoSlice(const int64_t& seqno)' uses the
address of 'seqno' in its definition.
The function breaks when a 'uint64_t' is passed in input, because
the implicit conversion from 'uint64_t' to 'int64_t' creates a
temporary, which is the actual argument passed to the function.
That causes a number of 'ep_testsuite' tests to fail on Jenkins
'kv_engine-linux-master-rocks' build.
This can happen because binding a rvalue to 'reference to const' is
legal in C++. Changing the function to take a 'const int64_t*' in input
fixes the issue and avoids any further misuse of a 'reference to const'
argument.
This change fixes the following 'ep_testsuite' tests:
  - ep_testsuite.value_eviction.rocksdb
  - ep_testsuite.full_eviction.rocksdb
  - ep_testsuite_basic.value_eviction.rocksdb
  - ep_testsuite_dcp.value_eviction.rocksdb
  - ep_testsuite_dcp.full_eviction.rocksdb

Change-Id: I5202cfd864de18b2d425d70626534dde03e39e0d
Reviewed-on: http://review.couchbase.org/84847
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>