Move numTotalItems from HashTable -> VBucket
In Full-Eviction, items may exist in a VBucket without being in the
HashTable, as they may have been fully evicted. As such, numTotalItems
is not a property of the HashTable, it is a property of the VBucket.

Therefore move numTotalItems from HashTable to VBucket, to better
encapsulate the VBucket's state.

Change-Id: I9d016fd45f393c4d678325471da429dfc1b6d0de
Reviewed-on: http://review.couchbase.org/84883
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>