[2.5/n] MB-25920: Fix a race condition in StaleItemDeleter test
In the previous commit, the test case that tests the pause-resume of
StaleItemDeleter task has a race condition wherein it tries to run
the HTCleaner task that is already run.

This commit fixes that by just running the next available task in
the future queue instead of trying to run 'HTCleaner task' particularly.

The test only needs to check if 'StaleItemDeleter task' is run multiple
times and if desired amount of items are purged, which it does.

Change-Id: I5a309a83b9a9205a5f017729ce8c4d67d4579c05
Reviewed-on: http://review.couchbase.org/84574
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>