MB-26476 Indicate the flusher task completed after updating stats
The 'seqno stats' test in the 'ep_testsuite' fails intermittently
when it checks the 'last_persisted_seqno' for a VBucket.
When we 'wait_for_flusher_to_settle', we wait for the
'EPStats.flusher_todo' to be equals to 0. But,
'KVBucket::flushVBucket()' updates the VBucket 'last_persisted_seqno'
after 'EPStats.flusher_todo' has been set to 0.
To fix, we set 'EPStats.flusher_todo = 0' only after that
'last_persisted_seqno' has been updated.

Change-Id: I2c0c98267d9b9997f692fb4b47b135289d3fbe17
Reviewed-on: http://review.couchbase.org/84683
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>