Refactor: Store the packet in the cookie
The cookie represents a single command so it should contain
the packet data for the actual command. Given that we (currently)
read all of the data into a shared buffer (possibly containing
multiple commands) we don't want to move the ownership of the
data into the cookie (that would mean having to split that
buffer).

Change-Id: I162e9ac6bb078326b45ac111df44efb7face4fcc
Reviewed-on: http://review.couchbase.org/84793
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>