MB-26255[5]: Process the final DCP item at response creation time
The item that is received from a backfill or from the checkpoint,
isn't the final item that is sent over the wire to the DCP consumer.
The item's value/xattrs is pruned or the value is compressed or
decompressed. So, before creating the final DCP response, create the
final item and store it in the response.

Change-Id: Icd96a54e89d4952e1099741c878d28e51eff1695
Reviewed-on: http://review.couchbase.org/84809
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>