MB-25342: Add Collections::VB::Manifest::CachingReadHandle
The caching readhandle allows limited read access to the manifest
but fits into the functional flow of front-end operations.

When constructing the CachingReadHandle from a key, the key is scanned
and a map lookup is performed. The result of the lookup is saved so
that further down the code path, subsequent isLogicallyDeleted checks
don't need to scan the key again and perform another map lookup.

Change-Id: Icffaf8a722f4a9e7e67bce870445cd3f75f940e3
Reviewed-on: http://review.couchbase.org/84474
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>