MB-25344: [1/n] Ignore logically deleted keys for ADD
The semantics of ADD mean that it should only succeed if the key does
not already exist. With collection deletion, the removal of keys is
'lazy', similar to expiry, thus it's possible for ADD to find a key
in the hash-table, which it can overwrite when it's logically deleted
or expired.

This change passes a CachingReadHandle down the ADD path (we already
had read access held on collections for the entire operation, so lock
scope is not changed here). Within the depths of ADD we can now
safely ignore logically deleted keys.

Change-Id: I9f30977474c2a292c7eb26f3529504d61e553429
Reviewed-on: http://review.couchbase.org/84310
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>