MB-25344: [2/n] Ignore logically deleted keys for REPLACE
The semantics of REPLACE mean that it should only succeed if the key
already exists. With collection deletion, the removal of keys is
'lazy', similar to expiry, thus it's possible for REPLACE to find a
key in the hash-table, which is actually logically deleted and should
trigger failure of the REPLACE.

This change passes a CachingReadHandle down the REPLACE path (we
already had read access held on collections for the entire operation,
so lock scope is not changed here). Within the depths of REPLACE we
can now safely work with logically deleted keys.

Change-Id: Iccc9c6370b7c6267ab4cc5b46baa63f9ccc64c8f
Reviewed-on: http://review.couchbase.org/84832
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>