MB-25344: [3/n] Ignore logically deleted keys for DELETE
The semantics of DELETE mean that it should fail if the request key
does not exist. With collection deletion, the removal of keys is
'lazy', similar to expiry, thus it's possible for DELETE to find a key
in the hash-table, which is actually logically deleted and should
trigger failure of the DELETE.

This change passes a CachingReadHandle down the DELETE path (we
already had read access held on collections for the entire operation,
so lock scope is not changed here). Within the depths of DELETE we can
now safely work with logically deleted keys.

Change-Id: I39ab8082ba26d08f8c885d73f775f22f0ba96595
Reviewed-on: http://review.couchbase.org/84833
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>