MB-25344: [11/n] Logically deleted keys and getMetaData
GetMetaData must ignore keys in deleted collections even if they may
hang around in the HT for some time. If collection deletion was
synchronous, then we would never be able to query a key in a deleted
collection, hence it's cleaner to say this key is gone rather than
return its meta.

Change-Id: I99ccfa2ff9fdf097d35f5d5cb1765c6dc3bdf129
Reviewed-on: http://review.couchbase.org/84841
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>