MB-22010: [1/N] Test for Hash Table Investigation
The test populates a hash table and then performs a sequence of gets
before evicting down to the low watermark.  The test is used to evaluate
the performance of the hash table eviction algorithm.

At the end of test it produces output stating for each document whether
it remained resident or was evicted.

The output can be fed into the accompanying visualisation script, which
produces a png file.  Each 10x10 pixel represents a document, and
indicates whether it remains in the hash table or has been evicted.

Note: The test is DISABLED as it is only to be run manually.

Change-Id: I194b7ebc018f0c807a38f3786e70a2a23c5a22be
Reviewed-on: http://review.couchbase.org/84931
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>