Add cb::mcbp::Header
The Header structure allows us to operate on the common fields
of a frame received from the network. The only two fields which
differentiate a Request from a Response is that the Request
carries a vbucket identifier whereas the Response carries a
status code.

We have cases in our code where we just want to inspect the
packet (and choose a different action if the packet is a
request or a response). In these cases it makes it more
readable if we can cast the buffer to a Header, and then
fetch the right structure off that.

Change-Id: I6e7da1ff81b141e9562e4e20f1a364b16d4ad068
Reviewed-on: http://review.couchbase.org/84932
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>