Refactor: let mcbp_collect_timings receive a cookie
When this method receives the cookie, it should also pick
out the opcode from the packet header, and not use the
cached version in the connection.

We should also not generate any timings for DCP connections,
as they use a different state diagram, so that the timings
are inaccurate. We need to properly solve that before adding
the timings back (currently the entire timer will reflect the
message being injected into the stream by step())

Change-Id: Id555ab3c1c06d3f6a5334162d4a7ffcaa480ef57
Reviewed-on: http://review.couchbase.org/85000
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>