MB-26619: RocksDB: put vbstate in same batch as updates
In saveDocs, we prepare a batch of updates to write to the DB,
followed by a persisted vbState, and then the batch is
committed. However the saveVBState is done using a non-batch Put(),
which means that the VBstate is persisted *before* the updates. This
is incorrect, as in the event of a crash we would see the wrong high
seqno etc in the vbState compared to the data.

Fix by moving the vbState update into the same batch, so it is
commited together.

Change-Id: I6f97e6f31161ca366e03229318eeb543e627797f
Reviewed-on: http://review.couchbase.org/84998
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>