MB-26521: RocksDB: split WriteBatches into write_buffer_size chunks
RocksDB's WriteBatch is atomic - all updates must be applied to a
single memtable to ensure crash-consistency. As a consequence, if a
WriteBatch exceeds the configured memtable size limit
(write_buffer_size) then the memtable must be expanded to be large
enough to contain the whole batch.

In RocksDBKVStore we use WriteBatch in saveDocs() when flushing a
VBucket to disk. Currently there's no limit (size or items) on how
much the flusher can flush at once. This means that if the DWQ grows
large (for example during a large bulk load), then the size of the
data flushed to disk - and consequently the size of the WriteBatch -
can significantly exceed the write_buffer_size.

For example, in the linked MB we saw the DWQ grow to ~40k items per
vBucket; storing 1kB items. That results in ~40MB of data being
written in a single batch, which is 40x larger than the
write_buffer_size. The effect of all of this is that significantly
more memory than expected was allocated for memtables - we expected
~256MB per node, where we saw 10GB. This resulted in the node going OOM.

To address this, split the commitBatch into write_buffer_size chunks -
if the batch grows too large, commit what we have and start a new
batch. This is functionally valid as we don't rollback at the batch
level, so there is no problem in splitting the batch into chunks.

Change-Id: Ie8c07ac418d20643d54639511cf7459cf36c7cb2
Reviewed-on: http://review.couchbase.org/85004
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>