MB-26603 [RocksDB] Allow to collect internal Statistics
RocksDB allow to collect cumulative performance Statistics over time.
With this change we allow to switch Statistics on and off depending on
the kv-engine configuration. We introduce a new 'rocksdb_stats_level'
parameter in 'configuration.json'. An empty value disables Statistics.
The periodic interval for RocksDB to collect Statistics can be set with
the 'stats_dump_period_sec' parameter in 'rockdb_options'.

Change-Id: I17111e18983d9b3fbc534fcca3606430201ba172
Reviewed-on: http://review.couchbase.org/85003
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>