MB-26681: RocksDB: Add config params for num background threads
Add support for specifying the number of RocksDB background threads
independently of the number of flusher / compaction tasks.

Flusher and Compaction tasks in RocksDB are per database (i.e
vBucket), not global - however we initially thought they were. As
such, the existing code to control the number of flushers/compactors
always sets the number of background threads to the same as the number
of tasks which run on those threads.

That would be fine if we had a single RocksDB database - we would have:

  - F flusher tasks running on F high priority threads, and
  - C compaction tasks runnning on C high priority threads.

However, given we have one DB per vBucket, this means we end up with:

  - F * V flusher tasks running on F threads, and
  - C * V compaction tasks running on C threads.

  Where V is the number of vBuckets.

As a result we cannot configure optimally in either direction - we
either end up with too many tasks scheduled across too many threads,
or too few tasks scheduled over too few threads.

This patch addresses this by making the (per-environment) thread
counts and the (per-DB) flusher and compaction tasks indpendent. It
adds two new config params:

* rocksdb_low_pri_background_threads
* rocksdb_high_pri_background_threads

In both instances a value of zero will auto-configure based on the
running system, or a non-zero value will override the default.

Change-Id: Ic72b7fe375e3c428d5e3f207bab08239d37f31dc
Reviewed-on: http://review.couchbase.org/85126
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>