MB-26631: Check if value requires modification at response creation
When a DCP response is created from a item from backfill or from the
checkpoint, the value or xattrs needs to be pruned or the item in
itself needs to be compressed/decompressed. A new item is created
in order to modify the value and then create a DCP response. It would
be pointless to create a copy of the item if the value doesn't require
any modification. Thus, create a copy of the item only if the value
needs modification.

Change-Id: I00b708e88d75fe6e35500f9cc8f68648ba111994
Reviewed-on: http://review.couchbase.org/85158
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>