MB-26705: Make ItemPager and associated tasks higher PRIO than DCP
DCP consumer has a task which is 'filling' memory, if we're running
out of memory, this filling task has a higher priority over the
ItemPager, so we may continue to fill memory whilst under memory
pressure.

Switch the priority so that the ItemPager would get scheduled ahead
of DCP so that we free memory first.

Change-Id: Ie74912386ef942fdbc9c1b906e91744e50cf581f
Reviewed-on: http://review.couchbase.org/85384
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>