Fix memory loss of a MockDcpConsumer obj created in a dcp_test
As observed in the recent failure of an address sanitizer test run,
we are not freeing the memory allocated to the MockDcpConsumer obj in
the test ConnectionTest.test_update_of_last_message_time_in_consumer().

This commit fixes it by using a RCPtr for the MockDcpConsumer obj.
Note: MockDcpConsumer obj is used as a shared obj using RCPtr at many
      places and hence we could not use a std::shared_ptr instead.

Change-Id: I12ad4693ae6fded3337a6cc919d4ea14dd48dbf7
Reviewed-on: http://review.couchbase.org/85230
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>