Reintroduce "Refactor: Remove c->binary_header (use the one in cookie)"
When moving to operating on multiple commands at the same time from
the client we can't use anything in the connection object to track
command state.

In the previous version of the patch we expected to be able to pick
up the opcode from the cookie if ewouldblock was set (during
disconnect from ep-engine). In some contexts there is no packet
available (for instance for DCP connections which is locked in the
ewouldblock state when they're waiting for a new mutation to
happen internally which it should send to the consumer). This
patch swallows the exception for non-existent packet and return
INVALID_OPCODE in these cases.

Change-Id: Id868d14991599060817ef2867fdeb7ce0a79e48c
Reviewed-on: http://review.couchbase.org/85236
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>