MB-26705: Make ItemPager's wake mechanism more reliable
The current mechanism which evaluates the task state is
racey and could lead to loss of a wakeup. It can happen
that the observer reads the state and see RUNNING but
the task is about to call snooze, thus the observer
will not wake the task.

This patch uses the atomic notify flag similar to DCP
checkpoint processor.

Change-Id: I2ff3e9c024474646cf470be41adef13872040549
Reviewed-on: http://review.couchbase.org/85385
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>