MB-26705: Re-run pager if memory is still high
Before the PagingVisitor finishes, re-check the status of memory.
We could still be in a memory condition situation and shouldn't
return without waking the ItemPager for another round.

test_memory_limit has been 'ported' to the single threaded model
as the original test became unstable for a number of reasons, primarily
that the ejection didn't free memory because the checkpoint manager
still had a reference on the Blob. The ported test has also removed
some of the final parts of the original test that were unstable.

Change-Id: I3a283beffba3418476d01000f6420ade1ab5fbd6
Reviewed-on: http://review.couchbase.org/85388
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>