MB-26470: Create 'Consumer Processor tsk' only before adding first stream
Currently we create the 'DCP Consumer Processor task' in the constructor
of the DCPConsumer obj. This is not fully safe because the DCPConsumer
passes a RCPtr of itself to the 'Processor task' even before the
consumer obj is fully created. If there is an exception in the
constructor then the 'Processor task' will be holding onto an
invalid reference. Further, it was also found that upon upgrading
RCPtr to std::shared_ptr, a std::bad_weak_ptr exception was thrown
if we tried to pass the reference ptr in the constructor.

Hence, this commit creates the 'Processor task' just before the
first stream is created on the consumer connection.

Change-Id: I414c32493772fbe83c2bb6cf718874e5b48f6a20
Reviewed-on: http://review.couchbase.org/85216
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>