GSL: Use not_null<> in server API
Tag non-null pointer args in server_api.h as gsl::not_null<>

GSL provides not_null<T> - this is a wrapper template which can be
applied to pointer types to indicate that pointer should never be
null. Instances of this type have two features:

    1. They cannot directly be assigned nullptr - this is enforced at
    compile-time.

    2. They throw an exception at runtime if they are ever assigned a
       value which happens to be null.

Note with (2), the exception is thrown at the point the variable is
assigned, not when it is later read; as is commonly the case in manual
null checks (for example in the head of a function).

In this patch, annotate all the function arguments in the Server API
which cannot be null with not_null<>. This highlights some
existing errors in unit tests, so fix these.

Change-Id: If7f9624d76b7c395abdb3121d7d7571008e8ccb6
Reviewed-on: http://review.couchbase.org/85378
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>