MB-26248: JSON datatype: Unconditionally validate user mutations
When mutating a value, always check if the value is JSON or not,
irrespective of the clients' JSON support.

When the Hello::Datatype flag was originally designed (3.0 timeframe),
it was proposed that the client (once Hello::JSON was negotiated)
would check if the item was JSON or not, and set request.datatype as
appropriate. However, this is potentially dangerous - if the client
incorrectly marks a non-JSON item as JSON, or a JSON-item as non-JSON,
then other parts of the system would fail to get the document they
expected - in other words there would be data loss / inaccessibility.

This is clearly undesirable, so we have decided to revisit this
policy, and instead *always* detect JSON on the server - essentially
we ignore whatever the client sets for the datatype.JSON field in
mutation requests.

Note while the code change in memcached is small; it causes a number
of tests to be updated; as we no longer care about what the client
sends for the JSON datatype flag.

Change-Id: I2d6d1dd87fe7c42e06e16a486b23a1129438f023
Reviewed-on: http://review.couchbase.org/85845
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>