MB-26791: Move read{Time,Size} histograms to kvtimings
The readTime and readSize statistics are histograms, which when
printed as stats are output as a sequence of start,end,count
fields. These arn't very easy to read in their raw form:

 ro_0:readTime_0,1:                             0
 ro_0:readTime_1,2:                             0
 ro_0:readTime_1024,2048:                       39387
 ro_0:readTime_1048576,2097152:                 77
 ro_0:readTime_1073741824,18446744073709551615: 0
 ro_0:readTime_128,256:                         13151
 ro_0:readTime_131072,262144:                   3099

Instead they are better rendered as histograms, which is what cbstats
does for all stats in the kvtimings group (which is where the
writeSize & Time histograms already live). Therefore move the read
stats to there.

Change-Id: I159b42d9906660206350988b07df8a4479d3c02c
Reviewed-on: http://review.couchbase.org/85517
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>