MB-26705: DCP should make disk items cold
DCP already transmits some 'nru' data in the mutation structure,
however this value is not a real cache heat because the Item created
for a SET (which eventually ends up in DCP sending) has an nru always
defaulted to INITIAL_NRU_VALUE (value of 2) and never 'copies' the
real nru from the cache. So in this commit we still send the nru of 2
for items being sent from memory but if an item came into the
DCP readyQueue from backfill only we make the mutation cold (i.e.
MAX_NRU_VALUE).

The consumer side now reads the NRU value from the incoming mutation
and constructs Items with that NRU value. So previosly we just used
the default value (2) for *all* mutations.

The result of this patch is that in-memory values still send nru:2 and
still get created on the consumer with nru:2, but now items from disk
have an nru:3 (cold). In a DGM rebalance this allows the pager to get
busy evicting the cold disk items on the first eviction pass rather
than doing multiple passes, the outcome is we keep mem_used away from
the quota and reduce the TMP_FAIL errors a new node may produce.

Change-Id: I2e0e54836edc7a917bb5d1d4c394b3ca79d5c4fe
Reviewed-on: http://review.couchbase.org/85649
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>