MB-26840: Re-order the deletion of EventuallyPersistentEngine members
In EventuallyPersistentEngine obj re-order when some of its members
are deleted. That is because in the ~EventuallyPersistentEngine(),
some of its members can call the others.

In this commit, dcpFlowControlManager_ outlives dcpConnMap_ and
kvBucket outlives dcpFlowControlManager_ and dcpConnMap_.

Also this commit converts the dcpFlowControlManager_ obj from
raw ptr to unique ptr.

Change-Id: Ic52fba66b6f92faf5b5364334055dcf1dd24363b
Reviewed-on: http://review.couchbase.org/85562
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>