MB-26849 [RocksDB] Avoid data races caused by 'delVBucket'
As we cannot create multiple RocksDB instances on the same DB,
RocksDBKVStore keeps a vector of handles for the opened DBs. Previously,
an handle was a unique_ptr. That allowed 'delVBucket' to release the
handle of a live DB (e.g., a DB used in a 'commit' execution).
With this change we model DB handles as shared_ptr, and we release a DB
handle in 'delVBucket' only when the thread executing the deletion is
the exclusive owner of the shared_ptr (i.e., when any other concurrent
operation on the same VBucket has completed).

Change-Id: Id9847083038ad7bae3796ede545710f03048229e
Reviewed-on: http://review.couchbase.org/85650
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>