MB-26470: Passive strm & Processor tsk to hold only a weak_ptr to consumer
Currently passive stream and Processor hold a shared_ptr to the
associated consumer conn and the consumer conn holds shared_ptr to them.
When a DcpConnMap is deleted without explicitly closing all its
connections, there is a cyclic reference causing memory leak.

Hence, this commit converts the shared_ptr (to consumer) held by
the passive stream and the 'processor task' to weak_ptr. It also adds a
test case that checks for the cyclic reference and results in a hang
as we will not be able to stop the executor pool task group which
contains the processor tsk.

Change-Id: I94562db9e110f1308af78dac1744bef5fe6ed414
Reviewed-on: http://review.couchbase.org/85714
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>