MB-26363: Send "DCP_STREAM_END" msg for a single stream close
As per the protocol, a DCP producer is supposed to send a "STREAM_END"
when a DCP stream is closed by the client. Currently due to an
implementation bug we send "STREAM_END" only when the producer initiates
the stream close, not when the client closes the stream.

Since this bug is present for a long time, some DCP clients might
have been conditioned not to receive any messages after the stream close.
So to avoid any backward compatibility issues we are adding a new
control message "send_stream_end_on_client_close_stream" which indicates
the DCP producer to send the "STREAM_END" message upon a stream close.
That is, if the DCP client sends this control message during the
connection set up, only then the producer sends the "STREAM_END" message.

Further, the control message "send_stream_end_on_client_close_stream" is
not recognised by the older versions of the DCP producer and hence the
clients can adapt accordingly by seeing the error while trying to set
the ctrl message.

Lastly, clients that do not use this control message will not
notice any change in the producer behavior.

Change-Id: I42b44bb710a8fc83be206db6d8bb698676ccc9ca
Reviewed-on: http://review.couchbase.org/84612
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>